Fase 3: Integrate PagedAttentionGpu.decode into AttentionLayer.forward
Route the attention softmax through the GPU decode kernel when CUDA is
available, falling back to the CPU reference path otherwise.

- hybrid_attn.zig: add PagedAttentionGpu field to AttentionLayer; init
  attempts cuda module load (graceful null if unavailable). forward() now
  branches: GPU path calls paged_gpu.decode() per token on the f16 memory
  pool; CPU path keeps the classic GQA-expanded softmax for validation.
- Add storeF16/loadF16 helpers; KV cache blocks now store K/V as f16
  (bytes_per_elem=2), matching the GPU kernel layout.
- main.zig: PagedKVCache dtype .f32 -> .f16 to match GPU blocks +
  llama.cpp convention. Test fixture dtype also .f16.

commit F3

+++ src/main.zig
@@ -320,7 +320,7 @@ fn runHybridInference(
         .head_dim = head_dim,
         .num_kv_heads = cfg.head_count_kv,
         .num_q_heads = cfg.head_count,
-        .dtype = .f32,
+        .dtype = .f16,
         .enable_prefix_cache = false,
         .enable_cpu_offload = false,
         .max_seq_len = max_seq_len,
